Fix bug in exit check
The first iteration had a bug where the only opportunity to return true
was if the current room required an item to exit. The default case was
always false, resulting in a Hotel California situation.

The fix is to check first if no items are needed (this is the hot path)
and return immediately. If not, only then do we check for a required
item. And if the required item is not inventory, only then do we return
false.