Teach parser to understand case insensitive names
Objects and features are already lower case, so this only applies to
room names. The player can use a lower case name, and the parser will
convert it to title case before parsing it.

> go upstairs hallway
You need to find an item before you can leave.
> look at rug
The rug is old and unraveling.
There is a thread you can pull, that might come in handy.
A long, loose thread from unraveling rug
> take thread
You have picked up the thread.
It is now in your inventory
> go to upstairs hallway
You are in the upstairs hallway.
When you exit the attic, you are on the south wall.
There are stairs in front of you, on the north wall.
To your left, on the west wall, is your parents' bedroom.
On the east wall there is your bedroom, and the bathroom.
Between your bedroom and the bathroom is a recycling bin filled to the top with seltzer cans.
Next to the door to your parents' room is an exercise ball and your mom's purse.
You need to look in the recycling bin to get the aluminum can.
You need to bounce on the exercise ball to open your parents' door.
You can look in your mom's purse, but you can't take anything out of it because she'll know you've been in there.

The recycling bin is full of seltzer cans.
Your mom's purse is gigantic. If you jump in you might get lost. Better not take anything or you might get caught.
The giant blue exercise ball that your dad uses instead of a desk chair.
>