Fix a logic bug in password entry
The player should not be able to enter the password or interact with the
computer until they climb the desk in the basement lab.

Previously, the computer was a hidden object that was discovered when the player
looked at the desk. But the computer should remain hidden until the player
climbs on the desk.

By removing the computer as a hidden object of the desk, we can change the
password function to check if the computer has been discovered. This works
because the default value of the computer's "discovered" field is "false", and
we can choose when it becomes "true".