Update the climb function
Currently, our climb function only works on desks. The player cannot
climb on other features, or even the rest of the furniture.

More importantly, in the basement lab the computer (which contains the
software needed to win the game) is not discovered until after the
player climbs on the desk.