Check for end game
At the end of each pass through the input loop we need to check if the current
room is now the attic and, if so, then check if the conditions have been met
to end the game.

Since we are also letting the player move around by just typing the room
name we will need to check for the same conditions in that branch as well.